[rutube] Move _extract_chapters_from_description from youtube.py to common.py
